requirements.txt
mediapipe>=0.10.0
opencv-python>=4.7.0
numpy>=1.24.0
pyserial
